Moves Roster to HipchatRoster, provides defaults for WILL_HANDLE, cleans things up.
