+++ will/mixins/__init__.py
@@ -2,7 +2,7 @@ from errors import ErrorMixin
 from email import EmailMixin
 from naturaltime import NaturalTimeMixin
 from room import RoomMixin
-from roster import RosterMixin
+# from roster import RosterMixin
 from plugins_library import PluginModulesLibraryMixin
 from schedule import ScheduleMixin
 from settings import SettingsMixin
